Small popup to accept/reject call.
You can now turn off popping up the whole Ant window on incoming call.
Instead, a small popup will be displayed at bottom right corner of the screen,
allowing you to accept or reject the incoming call (and, if you wish, turn on
the recording). You will see caller's number there.

The popup is on top, but doesn't receive focus, so you are not interrupted in
your work.